[feat] 루트 총 거리 int 값으로 타입 변경 및 valid 추가

+++ src/main/java/org/sopt/pawkey/backendapi/domain/routes/application/dto/command/RouteRegisterCommand.java
@@ -10,7 +10,7 @@ import lombok.Builder;
 @Builder
 public record RouteRegisterCommand(
 	List<Coordinate> coordinates,
-	Double distance,
+	int distance,
 	int duration,
 	LocalDateTime startedAt,
 	LocalDateTime endedAt,
